stats added

+++ Recuperation_donnees_CKAN/gql/schema.py
@@ -1,6 +1,15 @@
 from graphene_django import DjangoObjectType
+from graphql import GraphQLError
 from Recuperation.models import IDE, IMSE, SFR, Coordonnee, Etablissement, Regroupement
 import graphene
+import graphql_jwt
+from graphql_jwt.decorators import login_required
+
+#Classe pour l'authentification par token JWT
+class AuthenticateUser(graphene.ObjectType):
+	token_auth = graphql_jwt.ObtainJSONWebToken.Field()
+	verif_token = graphql_jwt.Verify.Field()
+	refresh_token = graphql_jwt.Refresh.Field()
 
 class CoordonneeType(DjangoObjectType):
 	class Meta:
@@ -12,10 +21,15 @@ class QueryCoordonnee(graphene.ObjectType):
 	all_coordonnees = graphene.List(CoordonneeType)
 	coordonnee_by_id = graphene.Field(CoordonneeType, 
 										id=graphene.Int(required=True))
-	
+	@login_required
 	def resolve_all_coordonnees(root, info):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		return Coordonnee.objects.all()
+	@login_required
 	def resolve_coordonnee_by_id(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			return Coordonnee.objects.get(pk=id)
 		except Coordonnee.DoesNotExist:
@@ -30,10 +44,15 @@ class QueryIMSE(graphene.ObjectType):
 	all_imse = graphene.List(IMSEType)
 	imse_by_id = graphene.Field(IMSEType, 
 										id=graphene.Int(required=True))
-
+	@login_required
 	def resolve_all_imse(root, info):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		return IMSE.objects.all()
+	@login_required
 	def resolve_imse_by_id(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			return IMSE.objects.get(pk=id)
 		except IMSE.DoesNotExist:
@@ -48,10 +67,15 @@ class QuerySFR(graphene.ObjectType):
 	all_sfr = graphene.List(SFRType)
 	sfr_by_id = graphene.Field(SFRType, 
 										id=graphene.Int(required=True))
-
+	@login_required
 	def resolve_all_sfr(root, info):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		return SFR.objects.all()
+	@login_required
 	def resolve_sfr_by_id(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			return SFR.objects.get(pk=id)
 		except SFR.DoesNotExist:
@@ -66,10 +90,15 @@ class QueryIDE(graphene.ObjectType):
 	all_ide = graphene.List(IDEType)
 	ide_by_id = graphene.Field(IDEType, 
 										id=graphene.Int(required=True))
-	
+	@login_required
 	def resolve_all_ide(root, info):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		return IDE.objects.all()
+	@login_required
 	def resolve_ide_by_id(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			return IDE.objects.get(pk=id)
 		except IDE.DoesNotExist:
@@ -85,10 +114,15 @@ class QueryRegroupement(graphene.ObjectType):
 	all_regroupements = graphene.List(RegroupementType)
 	regroupement_by_id = graphene.Field(RegroupementType, 
 										id=graphene.Int(required=True))
-	
+	@login_required
 	def resolve_all_regroupements(root, info):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		return Regroupement.objects.all()
+	@login_required
 	def resolve_regroupement_by_id(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			return Regroupement.objects.get(pk=id)
 		except Regroupement.DoesNotExist:
@@ -105,11 +139,15 @@ class QueryEtablissement(graphene.ObjectType):
 	all_etablissements = graphene.List(EtablissementType)
 	etablissements_by_id = graphene.Field(EtablissementType, 
 										id=graphene.Int(required=True))
-
+	@login_required
 	def resolve_all_etablissements(root, info):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		return Etablissement.objects.all()
-
+	@login_required
 	def resolve_etablissements_by_id(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			return Etablissement.objects.get(pk=id)
 		except Etablissement.DoesNotExist:
@@ -124,8 +162,10 @@ class CreateCoordonnee(graphene.Mutation):
 		telephone = graphene.String()
 
 	coordonnee = graphene.Field(CoordonneeType)
-
+	@login_required
 	def mutate(root, info, adresse, municipalite, code_postal, site=None, telephone=None):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		coordonnee = Coordonnee(
 			adresse=adresse,
 			municipalite=municipalite,
@@ -141,8 +181,10 @@ class DeleteCoordonnee(graphene.Mutation):
 		id = graphene.Int(required=True)
 
 	success = graphene.Boolean()
-
+	@login_required
 	def mutate(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			coordonnee = Coordonnee.objects.get(pk=id)
 			coordonnee.delete()
@@ -156,8 +198,10 @@ class CreateIDE(graphene.Mutation):
 		imse = graphene.Argument(graphene.ID, required=True)
 
 	ide = graphene.Field(IDEType)
-
+	@login_required
 	def mutate(root, info, sfr, imse, defavorisation):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		#Rechercher les objets liés (sfr, imse)
 		try:
 			sfr_obj = SFR.objects.get(pk=sfr)
@@ -180,8 +224,10 @@ class DeleteIDE(graphene.Mutation):
 		id = graphene.Int(required=True)
 
 	success = graphene.Boolean()
-
+	@login_required
 	def mutate(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			ide = IDE.objects.get(pk=id)
 			ide.delete()
@@ -195,8 +241,10 @@ class CreateIMSE(graphene.Mutation):
 		rang = graphene.Int(required=True)
 
 	imse = graphene.Field(IMSEType)
-
+	@login_required
 	def mutate(root, info, indice, rang):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		imse = IMSE(
 			indice=indice,
 			rang=rang
@@ -209,8 +257,10 @@ class DeleteIMSE(graphene.Mutation):
 		id = graphene.Int(required=True)
 
 	success = graphene.Boolean()
-
+	@login_required
 	def mutate(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			imse = IMSE.objects.get(pk=id)
 			imse.delete()
@@ -224,8 +274,10 @@ class CreateSFR(graphene.Mutation):
 		rang = graphene.Int(required=True)
 
 	sfr = graphene.Field(SFRType)
-
+	@login_required
 	def mutate(root, info, indice, rang):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		sfr = SFR(
 			indice=indice,
 			rang=rang
@@ -238,8 +290,10 @@ class DeleteSFR(graphene.Mutation):
 		id = graphene.Int(required=True)
 
 	success = graphene.Boolean()
-
+	@login_required
 	def mutate(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			sfr = SFR.objects.get(pk=id)
 			sfr.delete()
@@ -258,8 +312,10 @@ class CreateRegroupement(graphene.Mutation):
 		langue = graphene.String(required=True)
 
 	regroupement = graphene.Field(RegroupementType)
-
+	@login_required
 	def mutate(root, info, code, nom, nom_court, coordonnee, superficie, perimetre, langue):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		#Rechercher les objets liés (coordonnee)
 		try:
 			coordonnee_obj = Coordonnee.objects.get(pk=coordonnee)
@@ -283,8 +339,10 @@ class DeleteRegroupement(graphene.Mutation):
 		id = graphene.Int(required=True)
 
 	success = graphene.Boolean()
-
+	@login_required
 	def mutate(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			regroupement = Regroupement.objects.get(pk=id)
 			regroupement.delete()
@@ -308,10 +366,11 @@ class CreateEtablissement(graphene.Mutation):
 		regroupement = graphene.Argument(graphene.ID, required=True)
 
 	etablissement = graphene.Field(EtablissementType)
-
+	@login_required
 	def mutate(root, info, codeOrg, codeImm, nom, coordonnee, type, prescolaire, primaire, 
 			secondaire, professionnel, adulte, ide, regroupement):
-		
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		#Valider qu'un type est true
 		if not (prescolaire or primaire or secondaire or professionnel or adulte):
 			raise Exception("Veuillez sélectionner au moins un niveau d'éducation.")
@@ -349,8 +408,10 @@ class DeleteEtablissement(graphene.Mutation):
 		id = graphene.Int(required=True)
 
 	success = graphene.Boolean()
-
+	@login_required
 	def mutate(root, info, id):
+		if not info.context.user.is_authenticated:
+			raise GraphQLError("Authentification invalide. Veuillez vous connecter.")
 		try:
 			etablissement = Etablissement.objects.get(pk=id)
 			etablissement.delete()
@@ -362,7 +423,8 @@ class Query(QueryEtablissement, QueryCoordonnee, QueryIDE,
 			QueryIMSE, QuerySFR, QueryRegroupement, graphene.ObjectType):
 	test = graphene.String(default_value="Hello, GraphQL!") #query for testing endpoint
 
-class Mutation(graphene.ObjectType):
+#Besoin de l'authentification pour les mutations
+class Mutation(AuthenticateUser, graphene.ObjectType):
 	create_etablissement = CreateEtablissement.Field()
 	delete_etablissement = DeleteEtablissement.Field()
 	create_coordonnee = CreateCoordonnee.Field()
